Extract VisualEditor from AppContainer
