+++ src/components/html-text-input.js
@@ -112,10 +112,12 @@ export class HTMLInputView extends React.Component<PropsType, StateType> {
 export default compose( [
 	withSelect( ( select ) => {
 		const {
+			getEditedPostAttribute,
 			getEditedPostContent,
 		} = select( 'core/editor' );
 
 		return {
+			title: getEditedPostAttribute( 'title' ),
 			value: getEditedPostContent(),
 		};
 	} ),
@@ -123,6 +125,9 @@ export default compose( [
 		const { resetBlocks } = dispatch( 'core/block-editor' );
 		const { editPost } = dispatch( 'core/editor' );
 		return {
+			editTitle( title ) {
+				editPost( { title } );
+			},
 			onChange( content ) {
 				editPost( { content } );
 			},
